Reconfigure webpack to copy static assets
This update fixes the icon font not working in the builds. The error was caused by webpack failing to match the paths when generating the renderers.

It was resolved by removing font files from webpacking and instead copying a static directory into the rederer build folder using webpack copy plugin